feat: refactor extproc routers
