+++ semantic_router/pkg/extproc/processor.go
@@ -1,0 +1,83 @@
+package extproc
+
+import (
+	"log"
+
+	ext_proc "github.com/envoyproxy/go-control-plane/envoy/service/ext_proc/v3"
+)
+
+// Process implements the ext_proc calls
+func (r *OpenAIRouter) Process(stream ext_proc.ExternalProcessor_ProcessServer) error {
+	log.Println("Started processing a new request")
+	
+	// Initialize request context
+	ctx := &RequestContext{
+		Headers: make(map[string]string),
+	}
+
+	for {
+		req, err := stream.Recv()
+		if err != nil {
+			log.Printf("Error receiving request: %v", err)
+			return err
+		}
+
+		log.Printf("Processing message type: %T", req.Request)
+
+		switch v := req.Request.(type) {
+		case *ext_proc.ProcessingRequest_RequestHeaders:
+			response, err := r.handleRequestHeaders(v, ctx)
+			if err != nil {
+				return err
+			}
+			if err := sendResponse(stream, response, "header"); err != nil {
+				return err
+			}
+
+		case *ext_proc.ProcessingRequest_RequestBody:
+			response, err := r.handleRequestBody(v, ctx)
+			if err != nil {
+				return err
+			}
+			if err := sendResponse(stream, response, "body"); err != nil {
+				return err
+			}
+
+		case *ext_proc.ProcessingRequest_ResponseHeaders:
+			response, err := r.handleResponseHeaders(v)
+			if err != nil {
+				return err
+			}
+			if err := sendResponse(stream, response, "response header"); err != nil {
+				return err
+			}
+
+		case *ext_proc.ProcessingRequest_ResponseBody:
+			response, err := r.handleResponseBody(v, ctx)
+			if err != nil {
+				return err
+			}
+			if err := sendResponse(stream, response, "response body"); err != nil {
+				return err
+			}
+
+		default:
+			log.Printf("Unknown request type: %v", v)
+
+			// For unknown message types, create a body response with CONTINUE status
+			response := &ext_proc.ProcessingResponse{
+				Response: &ext_proc.ProcessingResponse_RequestBody{
+					RequestBody: &ext_proc.BodyResponse{
+						Response: &ext_proc.CommonResponse{
+							Status: ext_proc.CommonResponse_CONTINUE,
+						},
+					},
+				},
+			}
+
+			if err := sendResponse(stream, response, "unknown"); err != nil {
+				return err
+			}
+		}
+	}
+}
