+++ semantic_router/pkg/extproc/server.go
@@ -1,0 +1,83 @@
+package extproc
+
+import (
+	"fmt"
+	"log"
+	"net"
+	"os"
+	"os/signal"
+	"syscall"
+
+	ext_proc "github.com/envoyproxy/go-control-plane/envoy/service/ext_proc/v3"
+	"google.golang.org/grpc"
+)
+
+// Server represents a gRPC server for the Envoy ExtProc
+type Server struct {
+	router *OpenAIRouter
+	server *grpc.Server
+	port   int
+}
+
+// NewServer creates a new ExtProc gRPC server
+func NewServer(configPath string, port int) (*Server, error) {
+	router, err := NewOpenAIRouter(configPath)
+	if err != nil {
+		return nil, err
+	}
+
+	return &Server{
+		router: router,
+		port:   port,
+	}, nil
+}
+
+// Start starts the gRPC server
+func (s *Server) Start() error {
+	lis, err := net.Listen("tcp", fmt.Sprintf(":%d", s.port))
+	if err != nil {
+		return fmt.Errorf("failed to listen on port %d: %w", s.port, err)
+	}
+
+	s.server = grpc.NewServer()
+	ext_proc.RegisterExternalProcessorServer(s.server, s.router)
+
+	log.Printf("Starting LLM Router ExtProc server on port %d...", s.port)
+
+	// Run the server in a separate goroutine
+	serverErrCh := make(chan error, 1)
+	go func() {
+		if err := s.server.Serve(lis); err != nil && err != grpc.ErrServerStopped {
+			log.Printf("Server error: %v", err)
+			serverErrCh <- err
+		} else {
+			serverErrCh <- nil
+		}
+	}()
+
+	// Wait for interrupt signal to gracefully shut down the server
+	signalChan := make(chan os.Signal, 1)
+	signal.Notify(signalChan, syscall.SIGINT, syscall.SIGTERM)
+
+	// Wait for either server error or shutdown signal
+	select {
+	case err := <-serverErrCh:
+		if err != nil {
+			log.Printf("Server exited with error: %v", err)
+			return err
+		}
+	case <-signalChan:
+		log.Println("Received shutdown signal, gracefully stopping server...")
+	}
+
+	s.Stop()
+	return nil
+}
+
+// Stop stops the gRPC server
+func (s *Server) Stop() {
+	if s.server != nil {
+		s.server.GracefulStop()
+		log.Println("Server stopped")
+	}
+}
