+++ semantic_router/pkg/extproc/utils.go
@@ -1,0 +1,18 @@
+package extproc
+
+import (
+	"log"
+
+	ext_proc "github.com/envoyproxy/go-control-plane/envoy/service/ext_proc/v3"
+)
+
+// sendResponse sends a response with proper error handling and logging
+func sendResponse(stream ext_proc.ExternalProcessor_ProcessServer, response *ext_proc.ProcessingResponse, msgType string) error {
+	// log.Printf("Sending %s response: %+v", msgType, response)
+	if err := stream.Send(response); err != nil {
+		log.Printf("Error sending %s response: %v", msgType, err)
+		return err
+	}
+	log.Printf("Successfully sent %s response", msgType)
+	return nil
+}
